+++ src/main/java/com/github/johnmedlockdev/main/Main.java
@@ -1,35 +1,37 @@
 package com.github.johnmedlockdev.main;
 
 import com.github.johnmedlockdev.main.data.Data;
+import com.github.johnmedlockdev.main.database.Database;
 import com.github.johnmedlockdev.main.file.FileInfo;
 import com.github.johnmedlockdev.main.parse.ParseInput;
+
+import java.util.Scanner;
 
 public class Main {
     public static void main(String... args) {
 
-        ParseInput userInput = new ParseInput(args);
-        FileInfo fileInfo = new FileInfo(userInput);
-        Data data = new Data(userInput, fileInfo);
+        while (true) {
+            ParseInput userInput = new ParseInput(args);
 
-        switch (data.getMethod()) {
-            case "new":
-                data.createFile();
-                break;
-            case "add":
-                data.createInput();
-                break;
-            case "generate":
-                data.generate();
-                break;
-            case "predict":
-                data.getPrediction();
-                break;
-            default:
-                System.out.println("Not a valid method.");
-                break;
+            if (userInput.getMode().equals("")) {
+
+                FileInfo fileInfo = new FileInfo(userInput);
+                Data data = new Data(userInput, fileInfo);
+
+                data.logic();
+
+            } else {
+                if (userInput.getMode().equals("exit")) {
+                    break;
+                }
+                System.out.println("What's the name of the file you would like to input?");
+                Scanner scanner = new Scanner(System.in);  // Create a Scanner object
+                String fileName = scanner.nextLine();  // Read user input
+                Database dataDB = new Database(fileName);
+
+            }
+
         }
-
-
     }
 }