+++ src/main/java/com/github/johnmedlockdev/main/database/Database.java
@@ -1,0 +1,7 @@
+package com.github.johnmedlockdev.main.database;
+
+public class Database {
+    public Database(String fileName) {
+
+    }
+}