+++ src/main/java/com/github/johnmedlockdev/main/strategy/strategies/test2.java
@@ -5,10 +5,10 @@ import com.github.johnmedlockdev.main.strategy.Strategy;
 
 import java.util.Arrays;
 
-public class test extends Strategy {
+public class test2 extends Strategy {
     double[] values = super.processData();
 
-    public test(Data data) {
+    public test2(Data data) {
         super(data);
 
     }
@@ -16,7 +16,6 @@ public class test extends Strategy {
 
     public void advise() {
         System.out.println(Arrays.toString(values));
-
     }
 }